Add files via upload
This repository contains a smart contract for a basic banking system implemented in Solidity, designed to run on the Ethereum blockchain. The smart contract allows users to create accounts, deposit and withdraw Ether, and check their balances. It serves as an educational example of how decentralized finance (DeFi) applications can be developed using blockchain technology.